+++ .gitignore
@@ -162,3 +162,4 @@ cython_debug/
 .vscode/*
 *.drawio
 # tests/temp_data/*
+tempCodeRunnerFile.py
