+++ startrak-cl/__main__.py
@@ -1,0 +1,3 @@
+from _app.shell import ShellConsole
+
+app = ShellConsole()
