+++ startrak-cl/_types/alias.py
@@ -1,0 +1,3 @@
+from typing import Literal
+
+LanguageMode = Literal['py', 'sh', 'st']
