--- startrak-cl/tempCodeRunnerFile.py
@@ -1,1 +1,0 @@
-input_text
